+++ frontend/src/map/components/popups/boundaryDownload.vue
@@ -149,8 +149,6 @@ export default {
 			var ret = [];
 			ret.push({ caption: 'Texto con GeoJSON (.csv)', key: 'cg' });
 			ret.push({ caption: 'Texto con WKT (.csv)', key: 'cw' });
-			ret.push({ caption: 'Excel con GeoJSON (.xlsx)', key: 'xg' });
-			ret.push({ caption: 'Excel con WKT (.xlsx)', key: 'xw' });
 			ret.push({ caption: 'Shapefile (.shp)', key: 'hw' });
 			ret.push({ caption: 'GeoPackage (.gpkg)', key: 'gw' });
 			return ret;